requirements/requirements.txt
Package              Version
-------------------- -------
asgiref              3.8.1
Django               5.1.2
django-crum          0.7.9
django-widget-tweaks 1.5.0
pillow               11.0.0
pip                  24.2
psycopg2             2.9.10
sqlparse             0.5.1
tzdata               2024.2
